// software/Ionic/GeeksCarConfiguration/src/app/model/car-configuration.ts
export class CarConfiguration {
    version: number;
    maxSpeed: number;
    actuatorCenter: number;
    actuatorMin: number;
    actuatorMax: number;
    useJoystick: boolean;
    invertJoystickX: boolean;
    invertJoystickY: boolean;
    useRc: boolean;
    usePotGo: boolean;
    steeringPotCenter: number;
    steeringPotMin: number;
    steeringPotMax: number;
}